FIR2IR: refactor implicit cast insertion, part 1: implicit cast
